model-code/requirements.txt
accelerate==0.30.0
Bottleneck==1.3.7
certifi==2024.6.2
charset-normalizer==2.0.4
clip==1.0
filelock==3.13.1
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numexpr==2.8.7
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pip==23.3.1
psutil==5.9.8
pyprojroot==0.3.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.1
setuptools==68.2.2
six==1.16.0
sympy==1.12
threadpoolctl==2.2.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
wheel==0.43.0
xgboost==2.0.3
